tp2/requirements.txt
pymongo
python-dotenv
requests
colorama

